+++ packages/components-js/tests/vrt/specs/link.vrt.ts
@@ -61,6 +61,7 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
 
       await forceHoverState(page, '.hover > p-link >>> a');
       await forceHoverState(page, '.hover > p-link >>> span'); // with slotted <a>, the shadowed <span> is used for hover styling
+      await forceFocusState(page, '.focus > p-link'); // native outline should not be visible
       await forceFocusState(page, '.focus > p-link >>> a');
       await forceFocusState(page, '.focus > p-link a');
       await forceFocusHoverState(page, '.focus-hover > p-link >>> a');