+++ packages/components/src/components.d.ts
@@ -106,6 +106,7 @@ export namespace Components {
         "loading"?: boolean;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**
@@ -170,6 +171,7 @@ export namespace Components {
         "stretch"?: BreakpointCustomizable<boolean>;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**
@@ -756,6 +758,7 @@ export namespace Components {
         "stretch"?: BreakpointCustomizable<boolean>;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**
@@ -1362,6 +1365,7 @@ declare namespace LocalJSX {
         "loading"?: boolean;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**
@@ -1426,6 +1430,7 @@ declare namespace LocalJSX {
         "stretch"?: BreakpointCustomizable<boolean>;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**
@@ -2032,6 +2037,7 @@ declare namespace LocalJSX {
         "stretch"?: BreakpointCustomizable<boolean>;
         /**
           * To remove the element from tab order.
+          * @deprecated since v2.8.0, use tabindex="-1" instead
          */
         "tabbable"?: boolean;
         /**